Remove old GIN datasource
Same story as the previous commit:

The GinDataSource struct + methods were written to work with the old,
pre-GOGS GIN. The code was unused since then. The dsource and datasource
files are being merged into one.
