--- cmd/gindoid/dsource.go
@@ -1,333 +1,0 @@
-package main
-
-import (
-	"bytes"
-	"crypto/md5"
-	"crypto/rsa"
-	"crypto/x509"
-	"encoding/hex"
-	"encoding/pem"
-	"encoding/xml"
-	"fmt"
-	"io/ioutil"
-	"net/http"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"regexp"
-	"strings"
-	"time"
-
-	log "github.com/Sirupsen/logrus"
-	"golang.org/x/crypto/ssh"
-	"gopkg.in/yaml.v2"
-)
-
-type GinDataSource struct {
-	GinURL    string
-	GinGitURL string
-	pubKey    string
-}
-
-func (s *GinDataSource) getDOIFile(URI string, user OAuthIdentity) ([]byte, error) {
-	//git archive --remote=git://git.foo.com/project.git HEAD:path/to/directory filename
-	//https://github.com/go-yaml/yaml.git
-	//git@github.com:go-yaml/yaml.git
-	fetchRepoPath := ""
-	if splURI := strings.Split(URI, "/"); len(splURI) > 1 {
-		uname := strings.Split(splURI[0], ":")[1]
-		repo := splURI[1]
-		fetchRepoPath = fmt.Sprintf(GINREPODOIPATH, uname, repo)
-	} else {
-		return nil, nil
-	}
-	client := &http.Client{}
-	req, _ := http.NewRequest(http.MethodGet, fmt.Sprintf("%s%s", s.GinURL, fetchRepoPath), nil)
-	req.Header.Add("Auhoroisation", user.Token)
-	resp, err := client.Do(req)
-	if err != nil {
-		// todo Try to infer what went wrong
-		log.WithFields(log.Fields{
-			"path":   fetchRepoPath,
-			"source": DSOURCELOGPREFIX,
-			"error":  err,
-		}).Debug("Could not get doifile")
-		return nil, err
-	}
-	defer resp.Body.Close()
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"path":   fetchRepoPath,
-			"source": DSOURCELOGPREFIX,
-			"error":  err,
-		}).Debug("Could nort read from received doifile")
-		return nil, err
-	}
-	return body, nil
-}
-
-func (s *GinDataSource) Get(URI string, To string, key *rsa.PrivateKey) (string, error) {
-	ginURI := strings.Replace(URI, "master:", s.GinGitURL, 1)
-	log.WithFields(log.Fields{
-		"URI":     URI,
-		"gin_uri": ginURI,
-		"to":      To,
-		"source":  DSOURCELOGPREFIX,
-	}).Debug("Start cloning")
-
-	//Create tmp ssh keys files from the key provided
-	tmpDir, err := ioutil.TempDir("", "")
-	if err != nil {
-		log.WithFields(log.Fields{
-			"source": DSOURCELOGPREFIX,
-			"error":  err,
-		}).Error("SSH key tmp dir not created")
-		return "", err
-	}
-
-	cmd := exec.Command("git", "clone", "--depth", "1", ginURI, To)
-	env := os.Environ()
-	// If a key was provided we need to use that with nthe ssh
-	if key != nil {
-		_, privPath, err := WriteSSHKeyPair(tmpDir, key)
-		if err != nil {
-			log.WithFields(log.Fields{
-				"source": DSOURCELOGPREFIX,
-				"error":  err,
-			}).Error("SSH key storing failed")
-			return "", err
-		}
-		env = append(env, fmt.Sprintf("GIT_SSH_COMMAND=ssh -i %s", privPath))
-		cmd.Env = env
-	}
-	out, err := cmd.CombinedOutput()
-	log.WithFields(log.Fields{
-		"URI":     URI,
-		"gin_uri": ginURI,
-		"to":      To,
-		"out":     string(out),
-		"source":  DSOURCELOGPREFIX,
-	}).Debug("Done with cloning")
-	if err != nil {
-		log.WithFields(log.Fields{
-			"URI":     URI,
-			"gin_uri": ginURI,
-			"to":      To,
-			"source":  DSOURCELOGPREFIX,
-			"error":   string(out),
-		}).Debug("Cloning did not work")
-		return string(out), err
-	}
-	cmd = exec.Command("git", "annex", "sync", "--no-push", "--content")
-	cmd.Dir = To
-	cmd.Env = env
-	out, err = cmd.CombinedOutput()
-	if err != nil {
-		// Workaround for uninitilaizes git annexes (-> return nil)
-		// todo
-		log.WithFields(log.Fields{
-			"URI":     URI,
-			"gin_uri": ginURI,
-			"to":      To,
-			"source":  DSOURCELOGPREFIX,
-			"error":   string(out),
-		}).Debug("Repo was not annexed")
-		return string(out), nil
-	}
-	return string(out), nil
-}
-
-func (s *GinDataSource) MakeUUID(URI string, user OAuthIdentity) (string, error) {
-	fetchRepoPath := ""
-	if splURI := strings.Split(URI, "/"); len(splURI) > 1 {
-		uname := strings.Split(splURI[0], ":")[1]
-		repo := splURI[1]
-		fetchRepoPath = fmt.Sprintf("/users/%s/repos/%s/browse/master", uname, repo)
-	}
-	resp, err := http.Get(fmt.Sprintf("%s%s", s.GinURL, fetchRepoPath))
-	// todo error checking
-	if err != nil {
-		return "", err
-	}
-	bd, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return "", err
-	}
-	currMd5 := md5.Sum(bd)
-	return hex.EncodeToString(currMd5[:]), nil
-}
-
-// ValidDOIFile resturns true if the specified URI has a DOI file containing all necessary information.
-func (s *GinDataSource) ValidDOIFile(URI string, user OAuthIdentity) (bool, *DOIRegInfo) {
-	in, err := s.getDOIFile(URI, user)
-	if err != nil {
-		return false, nil
-	}
-	//Workaround as long as repo does spits out object type and size (and a zero termination...)
-	in = bytes.Replace(in, []byte("\x00"), []byte(""), -1)
-	re := regexp.MustCompile(`blob\W\d+`)
-	in = re.ReplaceAll(in, []byte(""))
-
-	doiInfo := DOIRegInfo{}
-	err = yaml.Unmarshal(in, &doiInfo)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"data":   string(in),
-			"source": DSOURCELOGPREFIX,
-			"error":  err,
-		}).Error("Could not unmarshal doifile")
-		return false, &DOIRegInfo{}
-	}
-	if !hasValues(&doiInfo) {
-		log.WithFields(log.Fields{
-			"data":    in,
-			"doiInfo": doiInfo,
-			"source":  DSOURCELOGPREFIX,
-			"error":   err,
-		}).Debug("DOI File misses entries")
-		return false, &doiInfo
-	}
-	return true, &doiInfo
-}
-
-type DOIRegInfo struct {
-	Missing     []string
-	DOI         string
-	UUID        string
-	FileSize    int64
-	Title       string
-	Authors     []Author
-	Description string
-	Keywords    []string
-	References  []Reference
-	Funding     []string
-	License     *License
-	DType       string
-}
-
-func (c *DOIRegInfo) GetType() string {
-	if c.DType != "" {
-		return c.DType
-	}
-	return "Dataset"
-}
-
-func (c *DOIRegInfo) GetCitation() string {
-	var authors string
-	for _, auth := range c.Authors {
-		if len(auth.FirstName) > 0 {
-			authors += fmt.Sprintf("%s %s, ", auth.LastName, string(auth.FirstName[0]))
-		} else {
-			authors += fmt.Sprintf("%s, ", auth.LastName)
-		}
-	}
-	return fmt.Sprintf("%s (%d) %s. G-Node. doi:%s", authors, time.Now().Year(), c.Title, c.DOI)
-}
-
-func (c *DOIRegInfo) EscXML(txt string) string {
-	buf := new(bytes.Buffer)
-	if err := xml.EscapeText(buf, []byte(txt)); err != nil {
-		log.Errorf("Could not escape:%s, %+v", txt, err)
-		return ""
-	}
-	return buf.String()
-
-}
-
-type Author struct {
-	FirstName   string
-	LastName    string
-	Affiliation string
-	ID          string
-}
-
-type NamedIdentifier struct {
-	URI    string
-	Scheme string
-	ID     string
-}
-
-func (c *Author) GetValidID() *NamedIdentifier {
-	if c.ID == "" {
-		return nil
-	}
-	if strings.Contains(strings.ToLower(c.ID), "orcid") {
-		// assume the orcid id is a four block number thing eg. 0000-0002-5947-9939
-		var re = regexp.MustCompile(`(\d+-\d+-\d+-\d+)`)
-		nid := string(re.Find([]byte(c.ID)))
-		return &NamedIdentifier{URI: "http://orcid.org/", Scheme: "ORCID", ID: nid}
-	}
-	return nil
-}
-func (a *Author) RenderAuthor() string {
-	auth := fmt.Sprintf("%s,%s;%s;%s", a.LastName, a.FirstName, a.Affiliation, a.ID)
-	return strings.TrimRight(auth, ";")
-}
-
-type Reference struct {
-	Reftype string
-	Name    string
-	DOI     string
-}
-
-type License struct {
-	Name string
-	URL  string
-}
-
-func hasValues(s *DOIRegInfo) bool {
-	if s.Title == "" {
-		s.Missing = append(s.Missing, MS_NOTITLE)
-	}
-	if len(s.Authors) == 0 {
-		s.Missing = append(s.Missing, MS_NOAUTHORS)
-	} else {
-		for _, auth := range s.Authors {
-			if auth.LastName == "" || auth.FirstName == "" {
-				s.Missing = append(s.Missing, MS_AUTHORWRONG)
-			}
-		}
-	}
-	if s.Description == "" {
-		s.Missing = append(s.Missing, MS_NODESC)
-	}
-	if s.License == nil || s.License.Name == "" || s.License.URL == "" {
-		s.Missing = append(s.Missing, MS_NOLIC)
-	}
-	if s.References != nil {
-		for _, ref := range s.References {
-			if ref.Name == "" || ref.Reftype == "" {
-				s.Missing = append(s.Missing, MS_REFERENCEWRONG)
-			}
-		}
-	}
-	return len(s.Missing) == 0
-}
-
-func WriteSSHKeyPair(path string, PrKey *rsa.PrivateKey) (string, string, error) {
-	// generate and write private key as PEM
-	privPath := filepath.Join(path, "id_rsa")
-	pubPath := filepath.Join(path, "id_rsa.pub")
-	privateKeyFile, err := os.Create(privPath)
-	defer privateKeyFile.Close()
-	if err != nil {
-		return "", "", err
-	}
-	privateKeyPEM := &pem.Block{Type: "RSA PRIVATE KEY", Bytes: x509.MarshalPKCS1PrivateKey(PrKey)}
-	if err = pem.Encode(privateKeyFile, privateKeyPEM); err != nil {
-		return "", "", err
-	}
-	privateKeyFile.Chmod(0600)
-	// generate and write public key
-	pub, err := ssh.NewPublicKey(&PrKey.PublicKey)
-	if err != nil {
-		return "", "", err
-	}
-	err = ioutil.WriteFile(pubPath, ssh.MarshalAuthorizedKey(pub), 0600)
-	if err != nil {
-		return "", "", err
-	}
-
-	return pubPath, privPath, nil
-}
